Minor upgrades

+++ robohive/envs/obs_vec_dict.py
@@ -5,8 +5,10 @@ Source  :: https://github.com/vikashplus/robohive
 License :: Under Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 ================================================= """
 
-import numpy as np
 from collections import deque
+
+import numpy as np
+
 
 class ObsVecDict():
     """
@@ -79,7 +81,7 @@ class ObsVecDict():
             self.initialize(obs_dict, ordered_obs_keys)
 
         # recover vec
-        obsvec = np.zeros(0)
+        obsvec = np.zeros(0, dtype=np.float32)
         for key in self.ordered_obs_keys:
             obsvec = np.concatenate([obsvec, obs_dict[key].ravel()]) # ravel helps with images
